catalognum: separate the functionality into its own class
